+++ NEWS.md
@@ -1,3 +1,9 @@
+# solvebio 2.12.0
+
+* Add support for enabling/disabling Global Beacon on datasets
+* Add ability to perform Global Search queries
+
+
 # solvebio 2.11.0
 
 * Fix issue with order fields from dataset and file queries